# backend/docs/swagger.yaml
definitions:
  github_com_observer_teatime_internal_auth.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  github_com_observer_teatime_internal_auth.RegisterInput:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
  github_com_observer_teatime_internal_domain.AttachmentDownloadResponse:
    properties:
      attachment_id:
        type: string
      download_url:
        type: string
      filename:
        type: string
      mime_type:
        type: string
      size_bytes:
        type: integer
    type: object
...
  github_com_observer_teatime_internal_domain.UploadCompleteRequest:
    properties:
      attachment_id:
        type: string
      sha256:
        type: string
    type: object
  github_com_observer_teatime_internal_domain.UploadInitRequest:
    properties:
      conversation_id:
        type: string
      filename:
        type: string
      mime_type:
        type: string
      size_bytes:
        type: integer
    type: object
  github_com_observer_teatime_internal_domain.UploadInitResponse:
    properties:
      attachment_id:
        type: string
      object_key:
        type: string
      presigned_url:
        type: string
      required_headers:
        additionalProperties:
          type: string
        type: object
    type: object
info:
  contact: {}
paths:
  /attachments/{id}/url:
    get:
      description: Get a presigned download URL for an attachment
      parameters:
      - description: Attachment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Download URL generated
          schema:
            $ref: '#/definitions/github_com_observer_teatime_internal_domain.AttachmentDownloadResponse'
        "403":
          description: Not authorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Attachment not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get file download URL
      tags:
      - attachments
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email/username and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_observer_teatime_internal_auth.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
...
  /auth/me:
    get:
      description: Get info about the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: string
              username:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get authenticated user
      tags:
      - auth
  /auth/refresh:
    post:
      description: Get a new access token using refresh token from cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              access_token:
                type: string
              expires_at:
                type: integer
              user: {}
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with username, email, and password
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_observer_teatime_internal_auth.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Username or email already exists
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
...
  /uploads/complete:
    post:
      consumes:
      - application/json
      description: Mark a file upload as complete after successfully uploading to
        R2
      parameters:
      - description: Upload completion request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_observer_teatime_internal_domain.UploadCompleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Upload completed
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Not authorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Attachment not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Complete file upload
      tags:
      - uploads
  /uploads/init:
    post:
      consumes:
      - application/json
      description: Request a presigned URL for uploading a file to R2
      parameters:
      - description: Upload initialization request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_observer_teatime_internal_domain.UploadInitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Presigned upload URL generated
          schema:
            $ref: '#/definitions/github_com_observer_teatime_internal_domain.UploadInitResponse'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Not a member of conversation
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Initialize file upload
      tags:
      - uploads
...
swagger: "2.0"